edit user profile, user infos can now be edited and updated without any issues. I made password readonly so, reset password on the home page will be the only way to change a password.

+++ SpringBootSigninSignupExample/src/main/java/com/utm/repository/UserRepository.java
@@ -1,13 +1,12 @@
 package com.utm.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import com.utm.model.User;
 
 @Repository("userReposistory")
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
 
 	User findByEmail(String email);
 }
